Store

+++ backend/src/main/java/com/backend/controller/store/StoreController.java
@@ -3,7 +3,10 @@ package com.backend.controller.store;
 import com.backend.domain.store.Store;
 import com.backend.service.store.StoreService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -17,11 +20,12 @@ public class StoreController {
     private final StoreService service;
 
     @PostMapping("/add")
-    public ResponseEntity add(Store store,
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity add(Authentication authentication, Store store,
                               @RequestParam(value = "files[]", required = false) MultipartFile[] files) {
 
         if (service.validate(store)) {
-            service.add(store, files);
+            service.add(store, files, authentication);
             return ResponseEntity.ok().build();
         } else {
             return ResponseEntity.badRequest().body("Validation failed");
@@ -30,6 +34,7 @@ public class StoreController {
 
     @GetMapping("list")
     public List<Store> list() {
+
         return service.list();
     }
 
@@ -41,13 +46,22 @@ public class StoreController {
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity delete(@PathVariable Integer id) {
-        service.remove(id);
-        return ResponseEntity.ok().build();
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity delete(@PathVariable Integer id, Authentication authentication) {
+
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Store store) {
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity edit(@RequestBody Store store, Authentication authentication) {
+        if (!service.hasAccess(store.getId(), authentication)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         service.edit(store);
         return ResponseEntity.ok().build();
     }
